pyproject.toml
[project]
name = "friendly-broccoli-discord"
version = "0.1.0"
description = "Telegram to Discord relay with AI translation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.33.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "telethon>=1.41.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/goob_ai"]
